scraper/requirements.txt
requests
neo4j
python-dotenv